+++ engine/src/main/java/dev/gamekit/components/RigidBody.java
@@ -13,7 +13,7 @@ import java.awt.*;
 import static dev.gamekit.utils.Math.*;
 
 /**
- * The {@link RigidBody} component enables physics-based motion for the entity.
+ * {@link RigidBody} enables physics-based motion for the entity.
  * <p>
  * During its {@link #update} phase and after a physics simulation step is complete, this
  * component will update the attached entity's {@link Transform} component's position and rotation