+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 - Added interpolation support to `DrawImage` draw call
 - Added `void drawImage(BufferedImage, int, int, int, int, ImageInterpolation)` method overload
   to `Renderer` class
+- Added `AnimatedSprite` which is a `Sprite` component extension that renders an animated sprite
+  sheet
 
 ### Changed